requirements.txt
Flask==2.1.1
psycopg2-binary
pandas==1.4.2
scikit-learn
numpy
setuptools==58.0.4
wheel==0.37.0
gunicorn==20.1.0

